Fix to #7348 - `Distinct()` after `Select()`ing a property doesn't generate correct SQL if query also contains optional navigation
Problem was that optional navigation is translated to GroupJoin which in turn forces materialization on entire entity. This causes some result operators (e.g. Distinct) to be applied on all columns of that entity, even if the customer specified a subset of columns to perform Distinct on.

Fix is to evaluate result operators on the client if their result depends on whether it was applied to a single column or entire row, and the query also contains GroupJoin.
Operators that can still be safely translated to SQL: Skip, Take, FirstOrDefault, Count, All, Any.